implemented another improvement from @mgrann03 to strongest lists
Now there is an option to show 'different type moves' on the strongest lists. This will allow pokemon with movesets containing different types to the selected one. Importantly, these types will count as not very effective towards the calculation of their moveset rating.

In the 'any type' strongest list, this option allows movesets whose fm and cm aren't matching. No changes on the effectivity of moves is made in this list since there isn't a 'selected type'.